v0.6.2 Cambio en servicio rol y usuario

+++ src/main/kotlin/utex/edu/e3/siasapi/Service/UsuarioService.kt
@@ -1,6 +1,7 @@
 package utex.edu.e3.siasapi.Service
 
 import org.springframework.beans.factory.annotation.Autowired
+import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
 import org.springframework.stereotype.Service
 import utex.edu.e3.siasapi.DTO.UsuarioDTO
@@ -9,10 +10,7 @@ import utex.edu.e3.siasapi.Repository.UsuarioRepository
 import java.util.*
 
 @Service
-class UsuarioService(
-    @Autowired private val usuarioRepository: UsuarioRepository,
-    @Autowired private val rolService: RolService
-    ) {
+class UsuarioService(@Autowired private val usuarioRepository: UsuarioRepository) {
 
     fun obtenerUsuarios(): List<UsuarioDTO> {
         val listaUsuario = usuarioRepository.findAll()
@@ -24,14 +22,15 @@ class UsuarioService(
         return listaUsuariosDTO
     }
 
-    fun obtenerUnUsuario(id:Long): UsuarioDTO {
+    fun obtenerUnUsuario(id:Long): ResponseEntity<UsuarioDTO> {
         val usuario: Optional<Usuario> = usuarioRepository.findById(id)
         val tempUsuario: Usuario
         if (usuario.isPresent) {
             tempUsuario = usuario.get()
-            return UsuarioDTO(tempUsuario.id,tempUsuario.correo,tempUsuario.estado,tempUsuario.ultimoLogin,tempUsuario.rol!!.rol)
+            val usuarioDTO = UsuarioDTO(tempUsuario.id,tempUsuario.correo,tempUsuario.estado,tempUsuario.ultimoLogin,tempUsuario.rol!!.rol)
+            return ResponseEntity.ok(usuarioDTO)
         } else {
-            return UsuarioDTO()
+            return ResponseEntity.notFound().build()
         }
     }
 
@@ -50,11 +49,13 @@ class UsuarioService(
         return UsuarioDTO(tempUsuario.id,tempUsuario.correo,tempUsuario.estado,tempUsuario.ultimoLogin,tempUsuario.rol!!.rol)
     }
 
-    //TODO Investigar porque no elimina el usuario
-    fun eliminarUsuario(id:Long) {
-        print(usuarioRepository.getById(id).correo)
-        usuarioRepository.deleteById(id)
-        print(usuarioRepository.existsById(id))
+    fun eliminarUsuario(id:Long): ResponseEntity<Boolean> {
+        if (usuarioRepository.existsById(id)) {
+            usuarioRepository.deleteById(id)
+            return ResponseEntity.ok(true)
+        } else {
+            return ResponseEntity.notFound().build()
+        }
     }
 
     fun existeUsuarioPorCorreo(correo:String):Boolean {
